pdf: implement outrotate
This requires the current transformation matrix to
be modified (with the 'cm' operator'), and that
uses trigonometric functions (which requires to be
linked with -lm). Since it changes the graphic context
we must now store it and then restore it, so 'q' before
'BT' and 'Q' after 'ET' (as equivalents to gsave
and grestore).

In PostScript, 'rotate' does more than just rotating
like what we do with 'cm': it also affects the
whole page. To achieve this in PDF we use the
/Rotate field in the /Page object.

The rotate degree was also changed to a double to
allow greater precision. This affects post, ps,
and pdf.

+++ post.h
@@ -56,7 +56,7 @@ void outrel(int h, int v);
 void outfont(int f);
 void outsize(int s);
 void outcolor(int c);
-void outrotate(int deg);
+void outrotate(double deg);
 void outeps(char *eps, int hwid, int vwid);
 void outpdf(char *pdf, int hwid, int vwid);
 void outlink(char *dst, int hwid, int vwid);
